refactor: re-pack inputs as PrelimOptions and SelvarsOptions
